[#54] Simplify `stdlib.Transactor`
