--- stdlib/entity.go
@@ -1,69 +1,0 @@
-package stdlib
-
-import (
-	"context"
-	"database/sql"
-
-	"github.com/kozmod/oniontx"
-)
-
-// Executor represents common methods of sql.DB and sql.Tx.
-type Executor interface {
-	Exec(query string, args ...any) (sql.Result, error)
-	ExecContext(ctx context.Context, query string, args ...any) (sql.Result, error)
-	Query(query string, args ...any) (*sql.Rows, error)
-	QueryContext(ctx context.Context, query string, args ...any) (*sql.Rows, error)
-	QueryRow(query string, args ...any) *sql.Row
-	QueryRowContext(ctx context.Context, query string, args ...any) *sql.Row
-	Prepare(query string) (*sql.Stmt, error)
-	PrepareContext(ctx context.Context, query string) (*sql.Stmt, error)
-}
-
-// TxOption implements oniontx.Option.
-type TxOption func(opt *sql.TxOptions)
-
-// Apply the TxOption to sql.TxOptions.
-func (r TxOption) Apply(opt *sql.TxOptions) {
-	r(opt)
-}
-
-// WithReadOnly set `ReadOnly` sql.TxOptions option.
-func WithReadOnly(readonly bool) oniontx.Option[*sql.TxOptions] {
-	return TxOption(func(opt *sql.TxOptions) {
-		opt.ReadOnly = readonly
-	})
-}
-
-// WithIsolationLevel set sql.TxOptions isolation level.
-func WithIsolationLevel(level int) oniontx.Option[*sql.TxOptions] {
-	return TxOption(func(opt *sql.TxOptions) {
-		opt.Isolation = sql.IsolationLevel(level)
-	})
-}
-
-// DB is sql.DB wrapper, implements oniontx.TxBeginner.
-type DB struct {
-	*sql.DB
-}
-
-func (db *DB) BeginTx(ctx context.Context, opts ...oniontx.Option[*sql.TxOptions]) (*Tx, error) {
-	var txOptions sql.TxOptions
-	for _, opt := range opts {
-		opt.Apply(&txOptions)
-	}
-	tx, err := db.DB.BeginTx(ctx, &txOptions)
-	return &Tx{Tx: tx}, err
-}
-
-// Tx is sql.Tx wrapper, implements oniontx.Tx.
-type Tx struct {
-	*sql.Tx
-}
-
-func (t *Tx) Rollback(_ context.Context) error {
-	return t.Tx.Rollback()
-}
-
-func (t *Tx) Commit(_ context.Context) error {
-	return t.Tx.Commit()
-}
